Adjusting the configuration
obamenu configuration has now been moved to ~/.config/openbox/obamenu.ini

This gives a per-user set of changes. If the file does not exist, obamenu will create one. If a value is deleted from the file, obamenu will use in-built defaults.

There's still an item of work to do to make some global settings, but at the moment I'm not sure that they are needed.